Updates to better serve separate pipelines; flushing out neighGraph and distMatrix

+++ src/pipePacket.hpp
@@ -10,8 +10,8 @@ class pipePacket {
 	struct pipeData{
 		std::vector<std::vector<double>> originalData;
 		std::vector<std::vector<double>> workingData;
-		std::vector<std::vector<int>> edges;
-		std::vector<int> weights;
+		std::vector<std::vector<unsigned>> edges;
+		std::vector<double> weights;
 	} workData;
 	
 };
